requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-ckeditor==6.5.0
django-js-asset==2.0.0
django-phonenumber-field==6.4.0
djangorestframework==3.13.1
idna==3.3
phonenumbers==8.12.54
Pillow==9.2.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
